+++ tests/entrypoints/openai/chat_completion/test_chat_completion.py
