+++ tests/entrypoints/openai/chat_completion/test_completion_with_image_embeds.py
