+++ tests/entrypoints/openai/completion/test_completion.py
